Extract composer dialog controller
