+++ assets/i18n/cht-hk.js
@@ -1,4 +1,4 @@
-import chtTwTranslations from './cht-tw.js?v=press-system-v3.4.41';
+import chtTwTranslations from './cht-tw.js?v=press-system-v3.4.42';
 
 export const languageMeta = { label: '繁體中文（香港）' };
