+++ assets/js/composer-markdown-drafts.js
@@ -4,7 +4,7 @@ import {
   hasMarkdownDraftContent,
   isEncryptedMarkdownDraftEntry,
   normalizeMarkdownContent
-} from './composer-markdown-state.js?v=press-system-v3.4.41';
+} from './composer-markdown-state.js?v=press-system-v3.4.42';
 
 const noop = () => {};
