+++ assets/js/publish/transports/connect-transport.js
@@ -1,4 +1,4 @@
-import { CONNECT_PUBLISH_MESSAGE_TYPE } from '../settings-store.js?v=press-system-v3.4.41';
+import { CONNECT_PUBLISH_MESSAGE_TYPE } from '../settings-store.js?v=press-system-v3.4.42';
 
 export function serializeConnectPublishFile(file) {
   const out = {
