+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.41",
-  "tag": "v3.4.41",
+  "version": "3.4.42",
+  "tag": "v3.4.42",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.40 <3.4.41"
+      ">=3.4.41 <3.4.42"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.40 first."
+    "message": "Update to v3.4.41 first."
   }
 }
