+++ scripts/test-system-release-package.sh
@@ -151,6 +151,11 @@ if ! grep -qx "press-system-${version}/assets/js/composer-notifications.js" "${e
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/composer-dialogs.js" "${entries_file}"; then
+  echo "expected package to include composer dialog overlay code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/composer-remote-sync.js" "${entries_file}"; then
   echo "expected package to include composer remote sync code" >&2
   exit 1
